Refactor build and source
