+++ engine/table/build.gradle
@@ -68,7 +68,6 @@ dependencies {
     compileOnly 'com.google.code.findbugs:jsr305:3.0.2'
 
     runtimeOnly project(':engine-tuplesource')
-    runtimeOnly project(':extensions-performance')
 
     Classpaths.inheritGuava(project)
