+++ extensions/performance/build.gradle
@@ -3,19 +3,10 @@ plugins {
     id 'java-library'
 }
 
-sourceSets {
-    main {
-        resources {
-            srcDir 'groovy'
-            include 'groovy/*.groovy'
-        }
-    }
-}
-
 dependencies {
     implementation project(':DHProcess')
     implementation project(':engine-table')
     implementation project(':Plot')
-
+    Classpaths.inheritAutoService(project)
     Classpaths.inheritGroovy(project, 'groovy', 'implementation')
 }
