+++ extensions/performance/src/main/resources/groovy/1-performance.groovy
