+++ server/build.gradle
@@ -10,6 +10,7 @@ dependencies {
     implementation project(':extensions-csv')
     implementation project(':extensions-arrow')
     implementation project(':extensions-parquet-table')
+    implementation project(':extensions-performance')
     implementation project(':extensions-jdbc')
     implementation project(':Util');
     implementation project(':Integrations')
