+++ server/jetty-app/build.gradle
@@ -10,6 +10,7 @@ configurations {
 dependencies {
     implementation project(':server-jetty')
 
+    runtimeOnly project(':extensions-performance')
     runtimeOnly project(':log-to-slf4j')
     runtimeOnly project(':logback-print-stream-globals')
     runtimeOnly project(':logback-logbuffer')
