+++ server/src/main/java/io/deephaven/server/console/groovy/InitScriptsModule.java
@@ -7,10 +7,10 @@ import dagger.Binds;
 import dagger.Module;
 import dagger.Provides;
 import dagger.multibindings.IntoSet;
+import io.deephaven.engine.table.impl.util.PerformanceQueries;
 import io.deephaven.engine.util.GroovyDeephavenSession.Base;
 import io.deephaven.engine.util.GroovyDeephavenSession.CountMetrics;
 import io.deephaven.engine.util.GroovyDeephavenSession.InitScript;
-import io.deephaven.engine.util.GroovyDeephavenSession.PerformanceQueries;
 import io.deephaven.engine.util.GroovyDeephavenSession.RunScripts;
 
 import java.util.Set;
@@ -29,7 +29,7 @@ public class InitScriptsModule {
 
         @Binds
         @IntoSet
-        InitScript bindsPerformanceQueriesScripts(PerformanceQueries impl);
+        InitScript bindsPerformanceQueriesScripts(PerformanceQueries.InitScript impl);
 
         @Provides
         static RunScripts providesRunScriptLogic(Set<InitScript> scripts) {
